Fix the bug in service Pre check handling
Previous commits while fixing the truthy issue of pre-check
introduced a bug in the handling of the environment variables.

In bash "cond1 && cond2" should be both true to have an exit value
of '0' since when the env is not set both become true since the
printing is successful as well. Which means that if the MINIO_VOLUMES
is not set the statement 'cond1 && cond2' was true i.e exit status '0'.

Once the MINIO_VOLUMES is set the statement 'cond1 && cond2' becomes
false i.e exit status '1'.

This is a tricky bash behavior to understand, to avoid this we simply
use explicit exit status and also use `if` conditions to provide
more clarity.